register resources and correct processes dir path
